esp8266: replacement of SDK library printf functions
In order to be able to control the log output from SDK libraries according to the defined log level, a number of library-specific printf functions were implemented in RIOT with PR 12998 for use by the SDK. These library-specific printf functions intercept the log output of the SDK and redirect it to the RIOT log macros.

To get these SDK library specific functions working, the printf/ets_printf calls in SDK libraries have to be replaced.